added, fetch helper and moved functionality to get pokemon list
