+++ src/App.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import logo from './img/logo-pokemon.png';
 import './App.css';
-import constants from './constants/index.js';
+import constants from './utilities/constants.js';
+import FetchHelper from './utilities/fetchHelper.js';
 
 import { SearchBar } from './components/SearchBar/index.jsx';
 import DetailView from './components/DetailView/index.jsx';
@@ -22,7 +23,6 @@ class App extends Component {
   }
 
   paginationClickHandler = (page_offset) => {
-    console.log("paginationClickHandler",page_offset)
     this.getPokemonList(page_offset)
   }
 
@@ -31,27 +31,29 @@ class App extends Component {
     this.setState({isEmptyState : false, profile : { "name": "pikachu", "id": "025", "img" : "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/25.png", "stats" :[{ "name": "hp", "base_stat": 35 },{ "name": "attack", "base_stat": 55 },{ "name": "defense", "base_stat": 40 },{ "name": "speed", "base_stat": 90 },{ "name": "Sp atk", "base_stat": 50 },{ "name": "Sp def", "base_stat": 50 }], "profile" :[{ "name" :"capture_rate", "value" : 190 },{ "name" :"egg_groups", "value" : "fairy, ground" },{ "name" :"abilities", "value" : "lightning-rod, static" },{ "name" :"gender_rate", "value" : 4 },{ "name" :"hatch_counter", "value" : 10 },{ "name" :"EVs", "value" : "0 Sp Att" }] }})
   }
 
-  getPokemonList = (offset) => {
-    const url = constants.pagination.url + "?limit=" + constants.pagination.limit + "&offset=" + (offset ? offset : 0);
-    fetch(url)
-      .then( res => res.json() )
-      .then( json => {
-        this.setState({
-          pokemons: json.results,
-          pages : {
-            previous:json.previous,
-            current_page: offset ? (offset/50) : 0,
-            next:json.next,
-            count: json.count
-          }
-        })
-      });
+  updatePokemonList = (pokemon_list, offset) => {
+    this.setState({
+      pokemons: pokemon_list.results,
+      pages : {
+        previous:pokemon_list.previous,
+        current_page: offset ? (offset/50) : 0,
+        next:pokemon_list.next,
+        count: pokemon_list.count
+      }
+    })
+  }
+
+  getPokemonList = (offset_param) => {
+    const url = constants.pagination.url;
+    const limit = constants.pagination.limit;
+    const offset = offset_param ? offset_param : 0
+    FetchHelper.get.pokemonList(offset, url, limit, this.updatePokemonList)
   }
 
   componentDidMount = () => {
     this.getPokemonList();
   }
-
+  
   render() {
     return (
       <div className="App">
